Add formatter attr_accessor to SyslogLogger
Rails 4 was failing to initialize in a staging/production environment
because ActiveSupport::TaggedLogging sets the formatter attribute of the
logger[1]. Since our SyslogLogger hasn't received love in a very long
time, its implementation has fallen out of date with the
ActiveSupport::Logger.

It is sufficient to just stick a formatter attribute, because
TaggedLogging does the management of its value and using it when
logging.

For what it's worth, Ruby 2.1.0 has a drop-in replacement syslog logger
that we will be able to use once we upgrade[2]. (Observant readers will
note that it exists in 2.0.0 too, but doesn't have support for logging
to multiple facilities, as we do to separate application traces from
EETS logs).

[1]: https://github.com/rails/rails/blob/1a9055073dcc189b289a1e3f47cbc72ff7380f0a/activesupport/lib/active_support/tagged_logging.rb#L57
[2]: http://ruby-doc.org/stdlib-2.1.0/libdoc/syslog/rdoc/Syslog/Logger.html

Change-Id: I2165a8e603a3fbfe0ee1ad5d5c92437d2c742cde
Reviewed-on: http://gerrit.causes.com/35714
Reviewed-by: Shane da Silva <shane@causes.com>
Tested-by: Tom Dooner <tom.dooner@causes.com>